Rename models to schemas
