--- backend/app/models/models.py
@@ -1,59 +1,0 @@
-# generated by fastapi-codegen:
-#   filename:  api.yml
-#   timestamp: 2025-05-19T12:37:04+00:00
-
-from __future__ import annotations
-
-from enum import Enum
-from typing import List, Optional
-
-from pydantic import BaseModel
-
-
-class Difficulty(Enum):
-    easy = "easy"
-    normal = "normal"
-    hard = "hard"
-
-
-class InterviewPostRequest(BaseModel):
-    source_code: Optional[bytes] = None
-    difficulty: Optional[Difficulty] = None
-    total_question: Optional[int] = None
-
-
-class InterviewPostResponse(BaseModel):
-    interview_id: Optional[str] = None
-    question: Optional[str] = None
-
-
-class InterviewPostResponse1(BaseModel):
-    error_message: Optional[str] = None
-
-
-class InterviewInterviewIdPostRequest(BaseModel):
-    question_id: Optional[int] = None
-    message: Optional[str] = None
-
-
-class InterviewInterviewIdPostResponse(BaseModel):
-    question_id: Optional[int] = None
-    response: Optional[str] = None
-
-
-class InterviewInterviewIdPostResponse1(BaseModel):
-    error_message: Optional[str] = None
-
-
-class InterviewInterviewIdGetResponse(BaseModel):
-    question_id: Optional[int] = None
-    question: Optional[str] = None
-
-
-class InterviewInterviewIdResultGetResponse(BaseModel):
-    scores: Optional[List[int]] = None
-    general_review: Optional[str] = None
-
-
-class InterviewInterviewIdResultGetResponse1(BaseModel):
-    error_message: Optional[str] = None
