+++ backend/app/schemas/schemas.py
@@ -1,0 +1,59 @@
+# generated by fastapi-codegen:
+#   filename:  api.yml
+#   timestamp: 2025-05-19T12:37:04+00:00
+
+from __future__ import annotations
+
+from enum import Enum
+from typing import List
+
+from pydantic import BaseModel
+
+
+class Difficulty(Enum):
+    easy = "easy"
+    normal = "normal"
+    hard = "hard"
+
+
+class InterviewPostRequest(BaseModel):
+    source_code: bytes
+    difficulty: Difficulty
+    total_question: int
+
+
+class InterviewPostResponse(BaseModel):
+    interview_id: str
+    question: str
+
+
+class InterviewPostErrorResponse(BaseModel):
+    error_message: str
+
+
+class InterviewInterviewIdPostRequest(BaseModel):
+    question_id: int
+    message: str
+
+
+class InterviewInterviewIdPostResponse(BaseModel):
+    question_id: int
+    response: str
+
+
+class InterviewInterviewIdPostErrorResponse(BaseModel):
+    error_message: str
+
+
+class InterviewInterviewIdGetResponse(BaseModel):
+    question_id: int
+    question: str
+
+
+class InterviewInterviewIdResultGetResponse(BaseModel):
+    scores: List[int]
+    general_review: str
+
+
+class InterviewInterviewIdResultGetErrorResponse(BaseModel):
+    error_message: str
